requirements.txt
torch
uvicorn~=0.34.0
fastapi~=0.115.11
pydantic~=2.10.6
python-pptx~=1.0.2
transformers~=4.49.0
diffusers~=0.32.2